// packages/cfdg/types/src/v2/qbo/http.ts
export type QboCustomerResponse = {
  /** Unique identifier of the customer */
  Id?: string;
  /** Reference to the parent customer, if any */
  ParentRef?: { value?: string };
  /** Display name of the customer */
  DisplayName?: string;
  /** Fully qualified name of the customer */
  FullyQualifiedName?: string;
  /** Billing address of the customer */
  BillAddr?: QboAddress;
  /** Shipping address of the customer */
  ShipAddr?: QboAddress;
  /** Whether the customer is active */
  Active?: boolean;
  /** Sync token of the customer */
  SyncToken?: string;
  /** Whether the customer is a job */
  Job?: boolean;
  /** Metadata of the customer including last updated time. */
  MetaData?: { LastUpdatedTime?: string };
};